Extract commonalities from Phone, Email and Address classes into Contact
